created ssa elimination module
